% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{cf_add_slides_ppt}
\alias{cf_add_slides_ppt}
\title{Add visualizations to Power Point slides}
\usage{
cf_add_slides_ppt(doc, slide_header, slide_plot, plt_width, plt_height)
}
\description{
Add visualizations to Power Point slides
}
\keyword{internal}
