% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-obs.R
\name{classroom_obs_add_ca}
\alias{classroom_obs_add_ca}
\title{Add Core Action (major and minor) identifiers}
\usage{
classroom_obs_add_ca(.data)
}
\arguments{
\item{.data}{The name of an R data frame containing classroom observations and placed into long form
with \code{tidy_forms_survey}.}
}
\value{
A data frame containg all the original columns and rows, along with two additional columns:
\code{core_action_main} and \code{core_action_minor}.
}
\description{
Classroom observations are based on the IPG and contain components called core actions.
There are major core actions, which are numbers, and minor core actions, which are letters.
THe IPG also contains Culture of Learning (CoL) and summary ratings. This function add two columns
to the data set of classroom observations identifying the major and minor core action.
}
\details{
The column \code{core_action_main} contains the major core action and \code{core_action_minor}
contains the minor core action.

The input data set should contain the classroom observations survey that has been placed into long
form with the function \code{tidy_forms_survey}. It must contain the \code{response_option} and
\code{question_stem} columns.
}
\examples{
tidy_forms_survey(classroom_observations_math, 8:ncol(classroom_observations_math), c(3,4,5)) |>
 classroom_obs_add_ca()

}
