% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-obs.R
\name{classroom_obs_add_tntpmetrics}
\alias{classroom_obs_add_tntpmetrics}
\title{Convert classroom observation data to a format usable by \code{tntpmetrics}}
\usage{
classroom_obs_add_tntpmetrics(.data, grade_column, subject_name, id_cols)
}
\arguments{
\item{.data}{The name of an R data frame containing classroom observations and placed into long form
with \code{tidy_forms_survey}, and containing the columns created by \code{classroom_obs_add_ca()}.}

\item{grade_column}{A string containing the column name containing the grade. For the grade column,
use 'K' or '0' for kindergarten. Use the whole number, as a string, for other grades.}

\item{subject_name}{The name of the subject. Must be one of: "Math", "Literacy", "Science", or "Social Studies".}

\item{id_cols}{A vector of strings signifying the columns to keep when transforming the dataset to wide-form. This correlates to the
\code{id_cols} parameter of \code{pivot_wider()}. \code{.id} will be added if it is not included.}
}
\value{
A data frame that can be used by \code{tntpmetrics} to calcualte IPG scores.
}
\description{
With the \code{tntpmetrics} package you can calculate IPG scores. This function puts observation
data in the proper format to be used by \code{tntpmetrics}. To use the function, take the raw
classroom observation data and first transform it with the \code{tidy_forms_survey()} function
and the \code{classroom_obs_add_ca()} function.
}
\examples{
tidy_forms_survey(classroom_observations_math, 8:ncol(classroom_observations_math), c(3,6)) |>
 classroom_obs_add_ca() |>
 classroom_obs_add_tntpmetrics(grade_column = 'grade', subject_name = 'Math',
                               id_cols = c('.id', 'when_did_the_observation_occur'))

}
