% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classroom_observations_math}
\alias{classroom_observations_math}
\title{Classroom observation pre and psot training.}
\format{
A data frame with 14 rows and 32 variables. Each row contains a teacher's observation
and is in the same format as Google Sheets creates for Google Forms.
}
\usage{
classroom_observations_math
}
\description{
A dataset containing example responses math classroom observations pre and post training
}
\keyword{datasets}
