% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cleaning.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Clean up column names from data sets imported from Google Forms surveys.}
\usage{
clean_column_names(column_names)
}
\arguments{
\item{column_names}{A string or vector of strings representing the column names we want to clean up.}
}
\description{
Clean up column names from data sets imported from Google Forms surveys.
}
\keyword{internal}
