% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-counts.R
\name{convert_df_to_list_single_site}
\alias{convert_df_to_list_single_site}
\title{Convert a data frame of parameters for a single site into a list}
\usage{
convert_df_to_list_single_site(single_df_parameters)
}
\arguments{
\item{single_df_parameters}{Data frame with parameters for a single site.
Must include the following columns: "site_name", "tool_name", "format",
"address", "name_column", "only_keep_distinct", "folder_url"}
}
\value{
A list where each row is an element in the list, names are column names,
and values are the values in the column for the given row.
}
\description{
Converts a data frame into a list of parameters for a single site that can be used by
functions to create response dashboards.
}
\keyword{internal}
