% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cleaning.R
\name{create_color_scales}
\alias{create_color_scales}
\title{Create a named vector from scales where the values are colors and names are the scales}
\usage{
create_color_scales(scale_name)
}
\arguments{
\item{scale_name}{name from \code{g2g_scale_order()}.}
}
\value{
A named vector with the scales as names and hex values as values.
}
\description{
To create manual fills in ggplot, you need a named vector with the colors as values and the scale names
as the names of the vector. This function takes a named vector where the scales values are values
colors are names, and switches names and values. The returned vector can then be used in ggplot
}
\examples{
create_color_scales('obs_yes_notyet')

}
