% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-counts.R
\name{create_or_return_sheet}
\alias{create_or_return_sheet}
\title{Return a Google Sheet or create a new one based on the Sheet's title}
\usage{
create_or_return_sheet(dashboard_title, folder_id)
}
\value{
Google Sheet that can be edited.
}
\description{
Given a Google Sheet's title, either return the sheet if it already exists
or create a new sheet with the given title if it does not exist. If more than
one sheet with the given title exists, return an error.
}
\keyword{internal}
