% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cleaning.R
\name{find_scale}
\alias{find_scale}
\title{Convert scale to a factor with levels in the proper order for plotting}
\usage{
find_scale(scale_column, use_agree_disagree)
}
\arguments{
\item{scale_column}{A vector with the responses in your data containing scales that you want to
convert to a factor.}

\item{use_agree_disagree}{For the strongly agree to strongly disagree scale, whether to use all
six points on the Liekrt scale or only agree and strongly agree.}
}
\value{
A vector with the same values as the input vector, \code{scale_column}, but converted to a
factor with the levels in the proper order
}
\description{
Given a vector of scales in your data and a list of possible scales, where each element in the
list is a different scale, the function will find the appropriate scale and convert the scale column
in your data to a factor with the levels in the proper order.
Find scale of common G2G questions
}
