% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{g2g_add_table_ppt}
\alias{g2g_add_table_ppt}
\title{Add tables to Power Point slides}
\usage{
g2g_add_table_ppt(
  doc,
  .data,
  slide_header,
  col_lengths,
  fontsize = 12,
  notes_text = NULL
)
}
\arguments{
\item{doc}{Document object. Created with \code{officer::read_pptx()}.}

\item{.data}{A data frame containing the table that will be added to a PPT slide.}

\item{slide_header}{The slide header, as a string.}

\item{col_lengths}{A numeric vector containing the length in inches of each column.
The vector's length should equal the number of columns.}

\item{fontsize}{The font size of column names and table text, as an integer. Defaults to 12.}

\item{notes_text}{Text, as a string, of notes to add to slide. Entire note must be one string.
Use \verb{\\n} within the string to add line breaks. Defaults to \code{NULL}, or no notes.}
}
\description{
Add tables to PPT slides. You must initialize the PPT deck with
\code{doc <- officer::read_pptx()}. After adding tables to the deck with this function, you can
write out the deck with \code{print(doc, target = 'file_name_of_deck.pptx')}
}
