% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{g2g_add_viz_ppt}
\alias{g2g_add_viz_ppt}
\title{Add visualizations to Power Point slides}
\usage{
g2g_add_viz_ppt(
  doc,
  slide_plot,
  slide_header,
  plt_width,
  plt_height,
  notes_text = NULL
)
}
\arguments{
\item{doc}{Document object. Created with \code{officer::read_pptx()}.}

\item{slide_plot}{A ggplot object containing the plot we want in the PPT slide.}

\item{slide_header}{The slide header, as a string.}

\item{plt_width}{The width of the plot, in inches, when it is in the PPT presentation.}

\item{plt_height}{The height of the plot, in inches, when it is in the PPT presentation.}

\item{notes_text}{Text, as a string, of notes to add to slide. Entire note must be one string.
Use \verb{\\n} within the string to add line breaks. Defaults to \code{NULL}, or no notes.}
}
\description{
Add ggplot plots to PPT slides as vector graphics. You must initialize the PPT deck with
\code{doc <- officer::read_pptx()}. After adding plots to the deck with this function, you can
write out the deck with \code{print(doc, target = 'file_name_of_deck.pptx')}
}
