% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cleaning.R
\name{g2g_aggregate_positive_responses}
\alias{g2g_aggregate_positive_responses}
\title{Sum percentage totals of positive responses}
\usage{
g2g_aggregate_positive_responses(
  .data,
  positive_responses,
  grouping_terms = NULL,
  only_keep_first_response = FALSE
)
}
\arguments{
\item{.data}{Input data frame. It must contain the aggregate percentage for each response option in long form.
For example, each row must be a question and response option combination (question 1, strongly agree).
It cannot be the raw data where each row is a response. The data set must be created with \code{g2g_forms_survey_calc_percentages()}.}

\item{positive_responses}{A string vector of positive responses, for example \code{c('Agree', 'Strongly Agree')}}

\item{grouping_terms}{Any columns that you want to group by, such as years, demographics, or pre or post training. Represented
as a vector of strings. Defaults to \code{NULL}, which means there are no groupings.}

\item{only_keep_first_response}{Boolean on whether the aggregate percentages should show up in all rows
or only the rows represented by the first response in \code{positive_responses}. Setting this to
\code{TRUE} is useful in stacked bar charts where you only want to show the percentages once and
have it mapped to one of the positive response options. Defaults to \code{FALSE}.}
}
\value{
The original data frame with an additional column titles \code{.strong_response_percent} that shows the
percentage of respondents who answered favorably (if the row represents a favorable response)
or unfavorable (if the row represents an unfavorable response).
}
\description{
We often want to sum the percentage of respondents answering favorably on a survey, for example
answering 'Agree' or 'Strongly Agree'. This function calculates these values and returns them as
a column in the original data. The returned data frame will have the same number of rows as the
input data frame, but it will have an additional column calls \code{.strong_response_percent}.
}
\examples{
results <- teacher_pre_survey |>
  g2g_tidy_forms_survey(8:30, 3) |>
  dplyr::mutate(assessment = 'Pre-survey') |>
  g2g_forms_survey_calc_percentages('assessment') |>
  # the function should only be used on a data set with common scales
  # so, filter to only keep a single question stem since we know it will have the same scales
  dplyr::filter(stringr::str_detect(.data[['question_stem']], 'To what extent do you agree'))

g2g_aggregate_positive_responses(results, c('Agree', 'Strongly Agree'), 'assessment')

}
