% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-obs.R
\name{g2g_calc_avg_ipg}
\alias{g2g_calc_avg_ipg}
\title{Calculate average IPG scores}
\usage{
g2g_calc_avg_ipg(.data, grade_column, subject_name, grouping_terms = NULL)
}
\arguments{
\item{.data}{Data set created with \code{g2g_tidy_forms_survey} and \code{g2g_classroom_obs_add_ca}}

\item{grade_column}{The column name, as a string, that contains the grade.}

\item{subject_name}{A single value, as a string, representing the subject for the observations.
Cannot have multiple subjects.}

\item{grouping_terms}{Column name or names, as a string or vector of strings, containing categories we want to group by when averaging
IPG scores. Each unique value in this column will be a distinct row in the final output.}
}
\value{
A data frame with average IPG scores. Each row is a different group identified in \code{grouping_terms}.
}
\description{
Calculates average IPG scores for identified groups. Data must be created with \code{g2g_tidy_forms_survey}
and \code{g2g_classroom_obs_add_ca}
}
\examples{
g2g_tidy_forms_survey(classroom_observations_math, 8:ncol(classroom_observations_math), c(3,6)) |>
    g2g_classroom_obs_add_ca() |>
    g2g_calc_avg_ipg(grade_column = 'grade', subject_name = 'Math',
    grouping_terms = 'when_did_the_observation_occur')

}
