% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-forms-survey.R
\name{g2g_calc_high_expectations}
\alias{g2g_calc_high_expectations}
\title{Create a data set that can be used by TNTP metrics to calculate high expectations.}
\usage{
g2g_calc_high_expectations(.data)
}
\arguments{
\item{.data}{A data frame containing the survey responses, created with \code{g2g_tidy_forms_survey}.}
}
\value{
A data frame that is in the correct format to be used by \code{tntpmetrics}.
}
\description{
The \code{tntpmetrics} package can be used to calculate high expectations metrics based on the common measures.
This function converts the data from the pre and post teacher surveys into the format required by \code{tntpmetrics}.
The input data must be created with \code{g2g_tidy_forms_survey}, or maintain the same form.
}
\examples{
teacher_pre_survey |>
   g2g_tidy_forms_survey(8:30, 3) |>
   g2g_calc_high_expectations()

}
