% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-forms-survey.R
\name{g2g_calc_high_expectations_averages}
\alias{g2g_calc_high_expectations_averages}
\title{Calculate average HE scores and percentage of teacher with HE for given group}
\usage{
g2g_calc_high_expectations_averages(.data, grouping_term = NULL)
}
\arguments{
\item{.data}{Data set returned from \code{g2g_tidy_forms_survey} and containing the four primary HE questions.}

\item{grouping_term}{Column name, as a string, to group by when calculating averages. Use \code{NULL}
(default) for no groups.}
}
\value{
A data frame with high expectations scores and the percentage of teachers with high expectations..
}
\description{
The \code{tntpmetrics} package can be used to calculate high expectations metrics based on the common measures and
the function \code{g2g_calc_high_expectations} is used to put data into a format accepted by \code{tntpmetrics}.
This function acts as a wrapper around those two functions by taking data created with \code{g2g_tidy_forms_survey},
using \code{g2g_calc_high_expectations} to put the data in the proper format for calculating high expectations,
calculating high expectations for each teacher with \code{tntpmetrics::make_metric('expectations')},
and then calculating average high expectations scores and the percentage of teachers with high
expectations per group.
}
\examples{
teacher_pre_survey |>
  g2g_tidy_forms_survey(8:30, 3) |>
  g2g_calc_high_expectations_averages()

}
