% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-forms-survey.R
\name{g2g_calc_inst_practices}
\alias{g2g_calc_inst_practices}
\title{Calculate a single score for all instructional practices questions}
\usage{
g2g_calc_inst_practices(.data, response_col, grouping_col = NULL)
}
\arguments{
\item{.data}{The data set, filtered to only include instructional practice questions.}

\item{response_col}{Column name, as string, containing scale responses.}

\item{grouping_col}{Column to group scores by.}
}
\value{
A data frame with the scores.
}
\description{
Calculate a single score for all instructional practices questions that is the average response,
after converting scales to numbers. The scale is converted to numbers with 0 being never and 5
being In All or Most Lessons.
}
