% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt-tools.R
\name{g2g_call_gpt}
\alias{g2g_call_gpt}
\title{Call GPT API}
\usage{
g2g_call_gpt(user_prompt, system_prompt, model = "gpt-4.1")
}
\arguments{
\item{user_prompt}{A character string containing the user prompt.}

\item{system_prompt}{A character string containing the system prompt.}

\item{model}{A character string representing the model to be used (default is "gpt-4o").}
}
\value{
API text response.
}
\description{
This function sends a request to the GPT API, passing the user and system prompts,
and returns the model's response. The function uses the API key located in \code{Sys.getenv("OPENAI_API_KEY")}.
}
\examples{
\dontrun{
g2g_user_prompt_survey <- g2g_user_prompt_survey("What did you like most about the training?", "It was very engaging.")
g2g_system_prompt_survey <- g2g_system_prompt_survey()
g2g_call_gpt(user_prompt, system_prompt)
}

}
