% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cleaning.R
\name{g2g_check_required_columns}
\alias{g2g_check_required_columns}
\title{Check for required columns}
\usage{
g2g_check_required_columns(.data, required_columns)
}
\arguments{
\item{.data}{The data set that we want to check for the required columns}

\item{required_columns}{A vector of strings representing required columns names}
}
\value{
Returns NULL if all columns are present. Otherwise, produces an error.
}
\description{
Check for required columns
}
