% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cleaning.R
\name{g2g_compare_names}
\alias{g2g_compare_names}
\title{Check participant names for consistency}
\usage{
g2g_compare_names(.data, participants, pre_post_col)
}
\arguments{
\item{.data}{A single data set containing pre and post data, in long form where pre and post are
in different rows.}

\item{participants}{The column name or names, as a string or vector of strings, that identify unique participants. This could
be names, emails or other identifiers.}

\item{pre_post_col}{The column name, as a string, that identifies whether a row is a pre or post
training observations.}
}
\value{
A dataframe containing unique pre and post participants, sorted so that users can manually
check for consistent spelling.
}
\description{
Analysis often require matching pre and post participants. An antecedent step is ensuring unique
identifiers such as names or email addresses are consistent across pre and post data sets.
This function returns a data set with unique pre and post participants, so that users can easily
manually check whether unique identifiers are consistent.
}
