% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-counts.R
\name{g2g_create_googlesheet_of_responses_from_list_single_site}
\alias{g2g_create_googlesheet_of_responses_from_list_single_site}
\title{Create dashboard on Google Sheets with responses}
\usage{
g2g_create_googlesheet_of_responses_from_list_single_site(
  list_of_tools,
  site_name,
  folder_url
)
}
\arguments{
\item{list_of_tools}{A list of all data tools. Each element in the list is a
separate tool. Each tool should also be a list with the following named values:
- 'tool_name': name of the tool ('Teacher Pre-Training Survey')
- 'format': Either 'Google Forms' or 'Qualtrics'
- 'address': For Google Forms, the web address to the Google Sheet containing
the results. For Qualtrics, the survey ID.
- 'name_column': The column name, as a string, where teacher names are located.
Here is an example of the list format:
list(
list(
tool_name = 'ELA Observations',
format = 'Google Forms',
address = "https://googlesheetsurl.com,
name_column = "Teacher's Last Name"
),
list(
tool_name = 'Student Survey',
format = 'Qualtrics',
address = "qualtricssurveyID",
name_column = "teacher_name"
)
)}

\item{site_name}{Name of site, as string. Added to sheet title and introductory
line of the first sheet in the dashboard.}

\item{folder_url}{URL, as a string, of the folder where you want to place the dashboard.
Folder must already be created.}
}
\value{
Displays dashboard in browser for checking. returns \code{NULL}.
}
\description{
Create a dashboard in Google Sheets with the number of responses for data tools.
Dashboard contains one sheet with total number of responses for each tool.
Additional sheets are also added that show which teachers responded for each tool.
Overwrites currently existing dashboards with the same name.
}
\keyword{internal}
