% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-counts.R
\name{g2g_create_googlesheet_response_dashboards}
\alias{g2g_create_googlesheet_response_dashboards}
\title{Create Google Sheet response dashboard for all sites in parameters data frame}
\usage{
g2g_create_googlesheet_response_dashboards(df_of_parameters)
}
\arguments{
\item{df_of_parameters}{Data frame with parameters for sites.
Must include the following columns:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ - "site_name": String of site name
 - "tool_name":" String of name of tool
 - "format": Format of the tool. Either 'Google Forms' or 'Qualtrics'
 - "address": url to responses Sheet for Google Forms, survey ID for Qualtrics
 - "name_column": Column in the data that contains teacher names
 - "only_keep_distinct": Boolean, whether to only keep distinct teacher names.
      Typically will be TRUE for students surveys, FALSE for others.
 - "folder_url": url to folder on Google Drive where we want the dashboard to live.
      Folder must already be created. The url must be the same for all distinct sites.
}\if{html}{\out{</div>}}}
}
\description{
Create dashboards in Google Sheets with the number of responses for data tools.
Each dashboard contains one sheet with total number of responses for each tool.
Additional sheets are also added that show which teachers responded for each tool.
Overwrites currently existing dashboards with the same name. One dashboard is
created for each unique value in the \code{site_name} column.
}
