% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_operations.R
\name{g2g_db_add_obs}
\alias{g2g_db_add_obs}
\title{Add observations to \code{observations} table}
\usage{
g2g_db_add_obs(con, .data)
}
\arguments{
\item{con}{Database connection to Good to Great database. Should be created with
\code{DBI::dbConnect(RSQLite::SQLite(), dbname = "good_to_great.db")}.}

\item{.data}{Dataframe of observations. Should be in long-form and must contain all required columns and no more.}
}
\value{
Message indicating success.
}
\description{
Add a dataframe of observations to the \code{observations} table.
}
