% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_operations.R
\name{g2g_db_add_site_info}
\alias{g2g_db_add_site_info}
\title{Add site information to \code{sites} table}
\usage{
g2g_db_add_site_info(con, site_info)
}
\arguments{
\item{con}{Database connection to Good to Great database. Should be created with
\code{DBI::dbConnect(RSQLite::SQLite(), dbname = "good_to_great.db")}.}

\item{site_info}{Site information list created with \code{g2g_site_information()}.}
}
\value{
Message indicating success.
}
\description{
Creates an additional row in the \code{sites} table that adds site information.
}
