% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_operations.R
\name{g2g_db_add_teacher_survey}
\alias{g2g_db_add_teacher_survey}
\title{Add teacher survey responses to \code{teacher_survey} table}
\usage{
g2g_db_add_teacher_survey(con, .data)
}
\arguments{
\item{con}{Database connection to Good to Great database. Should be created with
\code{DBI::dbConnect(RSQLite::SQLite(), dbname = "good_to_great.db")}.}

\item{.data}{Dataframe of teacher survey responses Should be in long-form and must contain all required columns and no more.}
}
\value{
Message indicating success.
}
\description{
Add a dataframe of survey responses to the \code{teacher_survey} table.
}
