% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_operations.R
\name{g2g_db_drop_site_data_from_table}
\alias{g2g_db_drop_site_data_from_table}
\title{Drop all site data from a single table}
\usage{
g2g_db_drop_site_data_from_table(con, site_id, db_table)
}
\arguments{
\item{con}{Database connection to Good to Great database. Should be created with
\code{DBI::dbConnect(RSQLite::SQLite(), dbname = "good_to_great.db")}.}

\item{site_id}{An integer. The value in the \code{site_id} column from the \code{sites} table.
Can be found with \code{g2g_site_information()}.}

\item{db_table}{String identifying which table we want to delete data from.
One of 'teacher_information', 'observations', 'teacher_survey', or 'student_survey'.}
}
\description{
Drops all data from a single site within a single table. Operation is irreversible
and it is recommended that you make a copy of your data base prior to dropping site data.
}
\details{
To drop data from the \code{teacher_information} table you must first drop the data from
the \code{observations}, \code{teacher_survey}, and \code{student_survey} tables. This is because
these tables contain foreign keys that rely on the \code{teacher_information} table.
}
