% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_operations.R
\name{g2g_db_get_site_info_key}
\alias{g2g_db_get_site_info_key}
\title{Get \code{site_id} (primary key for site) from \code{sites} table}
\usage{
g2g_db_get_site_info_key(con, site_info)
}
\arguments{
\item{con}{Database connection to Good to Great database. Should be created with
\code{DBI::dbConnect(RSQLite::SQLite(), dbname = "good_to_great.db")}.}

\item{site_info}{Site information list created with \code{g2g_site_information()}.}
}
\value{
Single integer representing primary key value for site.
}
\description{
Get the primary key value from a site so that it can be used when adding data.
}
