% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-obs.R
\name{g2g_first_or_last}
\alias{g2g_first_or_last}
\title{Find the first and last observations for a person, based on the date.}
\usage{
g2g_first_or_last(.data, grouping_columns, date_column)
}
\arguments{
\item{.data}{Data frame containing observation data.}

\item{grouping_columns}{Columns to group by when finding the first and last observations, as
a vector of strings. Can be names or other unique identifiers.}

\item{date_column}{Column containing dates that will be used to determine the first and last.}
}
\value{
A vector of strings with the same length as the number of rows in \code{.data} containing
whether the observation is the first, last, or in the middle of the dates.
}
\description{
Find the first and last observations for a person, based on the date.
}
