% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cleaning.R
\name{g2g_id_pre_post}
\alias{g2g_id_pre_post}
\title{Identify participants who have values in pre and post tools}
\usage{
g2g_id_pre_post(.data, participants, pre_post_col)
}
\arguments{
\item{.data}{A single data set containing pre and post data, in long form where pre and post are
in different rows.}

\item{participants}{The column name, as a string, that identifies unique participants. This could
be names, emails or other identifiers.}

\item{pre_post_col}{The column name, as a string, that identifies whether a row is a pre or post
training observations.}
}
\value{
A vector of values from the \code{participants} column that includes participants in the
pre and post data sets.
}
\description{
Analysis often require comparing pre and post metrics, only using participants in both the
pre and post datasets. For example, we might calculate changes in high expectations scores
between pre and post, only using teachers who took pre and post surveys. This function identifies
participants with pre and post observations and returns a vector with their names / emails / id numbers.
}
