% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cleaning.R
\name{g2g_list_of_scales}
\alias{g2g_list_of_scales}
\title{Named list of scales used in Good to Great}
\usage{
g2g_list_of_scales()
}
\description{
Named list of scales used in Good to Great
}
\keyword{internal}
