% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{g2g_notes_from_data}
\alias{g2g_notes_from_data}
\title{Create notes based on items in your data}
\usage{
g2g_notes_from_data(.data, note_category, note_description, note_title = NULL)
}
\arguments{
\item{.data}{A data frame containing columns with the categories and descriptions.}

\item{note_category}{Column name, as a string, of the column containing the category (CA1a, CA1b, etc.)}

\item{note_description}{Column name, as a string, of the column containing the description of the category.}

\item{note_title}{A string that will be placed on the top of the note.}
}
\value{
A string with the note, which can be added to a slide with the \code{notes_text} parameter of
\code{g2g_add_viz_ppt} or \code{g2g_add_table_ppt}.
}
\description{
Your data might contain information that is useful for notes. For example, one column might contain
categories (CA1a, CA1b, etc.) and a different column might contain descriptions of the category.
This function creates notes from the categories and descriptions. it is of the format, "category: description".
}
