% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teacher-survey.R
\name{g2g_number_times_teacher_answered}
\alias{g2g_number_times_teacher_answered}
\title{Identify how many times a teacher answered a question}
\usage{
g2g_number_times_teacher_answered(.data)
}
\arguments{
\item{.data}{Data frame with pre and post-training data. created from \code{g2g_teacher_combine_pre_post()}}
}
\value{
the same data frame, with an additional column called \code{n_question_answers} that is an integer telling
how many times a teacher answered a question in total.
}
\description{
Helps identify how many times the teacher answered a question, which aids in identifying teachers
who did not answer a pre and post-training question on both the pre and post-training surveys.
Teachers are identfied by their email address in the \code{email} column.
}
