% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-obs.R
\name{g2g_obs_add_viz_ppt}
\alias{g2g_obs_add_viz_ppt}
\title{Add an observation visualization to a PPT slide}
\usage{
g2g_obs_add_viz_ppt(
  doc,
  plt,
  ca_descriptions,
  core_action,
  plt_height = 6,
  plt_width = 9
)
}
\arguments{
\item{doc}{The PPT document object. Initially created with \code{g2g_create_deck_ppt()}.}

\item{plt}{The plot to add to the PPT. Typically created with \code{g2g_obs_create_viz()}.}

\item{ca_descriptions}{A description of all core actions. Can be created with \code{.data |> distinct(core_action_main, core_action_minor, core_action, response_option)}}

\item{core_action}{The core action that we want to get data for. A string that mirrors the spelling in the
\code{core_action_main} column.}

\item{plt_height}{The height of the plot in the PPT. Can be calculated with \code{g2g_ppt_calculate_plot_height()}.
Defaults to 6.}

\item{plt_width}{The width of the plot. Defaults to 9.}
}
\value{
A \code{doc} object that can be reused when adding additional slides to the PPT.
}
\description{
Add an observation visualization to a PPT slide
}
