% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-obs.R
\name{g2g_obs_create_viz}
\alias{g2g_obs_create_viz}
\title{Create visualization of observation data.}
\usage{
g2g_obs_create_viz(.data, core_action)
}
\arguments{
\item{.data}{Data created with \code{g2g_obs_get_ca_data()}}

\item{core_action}{The core action that we want to get data for. A string that mirrors the spelling in the
\code{core_action_main} column.}
}
\description{
Creates a horizontal bar plot for a single Core Action. Each sub-Core Action is on a seperate row
(y-axis) of the bar chart.
}
