% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-obs.R
\name{g2g_obs_long_form}
\alias{g2g_obs_long_form}
\title{Place observation data into long form and add columns identifying the core action}
\usage{
g2g_obs_long_form(.data, ...)
}
\arguments{
\item{.data}{The raw observation data, as pulled from Google Sheets}

\item{...}{for \code{g2g_tidy_forms_survey}. Should include the \code{question_columns} and
\code{grouping_columns} parameters}
}
\value{
A tibble of the observation data in long form, where each row is a teacher / core action combination.
}
\description{
Place observation data into long form and add columns identifying the core action
}
