% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-obs.R
\name{g2g_obs_map_scales}
\alias{g2g_obs_map_scales}
\title{Find the scales for observations based on the Core Action}
\usage{
g2g_obs_map_scales(core_action)
}
\arguments{
\item{core_action}{The core action whose scale is needed. String.
Options are '1', '2', '3', 'RFS', 'Demands of the Standards', 'Culture of Learning'.}
}
\value{
A vector with the scales for the core action, in order
}
\description{
Find the scales for observations based on the Core Action
}
