% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-obs.R
\name{g2g_obs_test_results}
\alias{g2g_obs_test_results}
\title{Test to ensure the aggregated data matches the  raw data}
\usage{
g2g_obs_test_results(
  raw_data,
  long_form_data,
  aggregated_data,
  core_action_to_test
)
}
\arguments{
\item{raw_data}{The raw data, as imported from Google Sheets}

\item{long_form_data}{The long form data of each individual response, created with \code{g2g_obs_long_form()}}

\item{aggregated_data}{Aggregated data showing the percentage of respondents for each question / Core Action,
created with \code{g2g_obs_calc_perc()}.}

\item{core_action_to_test}{The Core Action to test. Will be the string in the \code{core_action} column of
the long form data. But, leave off "CA ".}
}
\value{
Returns NULL if test is passed. otherwise, returns \code{testthat} output.
}
\description{
Test to ensure the aggregated data matches the  raw data
}
