% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-obs.R
\name{g2g_obs_viz_overall}
\alias{g2g_obs_viz_overall}
\title{Create single plot that contains all overall observation scores}
\usage{
g2g_obs_viz_overall(
  .data,
  height_relationships = c(1, 3),
  first_obs_factor = "First Observation"
)
}
\arguments{
\item{.data}{Dataset containg aggregate overall Core Action values. Created by first calculating
aggregates with \code{g2g_obs_calc_perc()} and then filtering to only keep overall scores.}

\item{height_relationships}{Relationship in proportions between plots with different scales, as a numeric vector.
Defaults to c(1, 3).}

\item{first_obs_factor}{String representing the first observation category in \code{.timing}, as a string.
\code{.timing} will be changed to a factor with this string being the first level. Useful for ordering plots.
Defaults to 'First Observation'.}
}
\value{
a ggplot plot
}
\description{
The input dataset \code{.data} must be transformed to where the values in the \code{core_action_main} column
represent unique scales. For example, Core Action 1 has different scales than Core Actions 2 and 3.
Core Actions 2 and 3 have the same scales. Therefore, \code{core_action_main} must be transformed to
where Core Action 1 has a value (ex: '1') and Core Actions 2 and 3 have the same value (ex: '2').
This value should represent the core action value needed for the scales. See \code{examples} for more information.
}
\examples{
\dontrun{

# regular expression to identify the overall items in the core_action column
overall_items <- c("Overall|Culture|Demands|RFS")

overall_scores_ca <- obs_percent |>
  # only keep the overall core action items
  filter(str_detect(core_action, !!overall_items)) |>
  # change `core_action_main` values so that core actions with the same scales have the same values,
  # and these values mirror the scale
  mutate(core_action_main = ifelse(core_action_main == "1", '1', "2"))

# create plot
g2g_obs_viz_overall(overall_scores_ca)

}

}
