% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{g2g_ppt_calculate_plot_height}
\alias{g2g_ppt_calculate_plot_height}
\title{Calculate optimal height for a visualizations that is being inserted into a PPT.}
\usage{
g2g_ppt_calculate_plot_height(column_with_rows)
}
\arguments{
\item{column_with_rows}{The column, as a vector, of your data that will be the rows in the
horizontal bar chart. The number of rows should equal the number of unique items in this column.}
}
\description{
This function is useful for horizontal bar charts. It calculates an optimal height for the plot
when it is placed into the PPT based on the number of rows in the horizontal bar chart.
}
