% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cleaning.R
\name{g2g_pull_data}
\alias{g2g_pull_data}
\title{Pull data from Google Sheets and add descriptive columns}
\usage{
g2g_pull_data(url, tool_name, site_info)
}
\arguments{
\item{url}{The url to the Google sheet}

\item{tool_name}{The name of the tool ('Observations', 'Post-Training')}

\item{site_info}{A list created by \code{g2g_site_information()} containing site information.}
}
\value{
A tibble with the data from the Google Sheet and additional columns called: \code{site_name}, \code{semester},
\code{subject}, and \code{tool}
}
\description{
Pulls data from Google Forms and adds columns to the data signifying the site, semester, subject, and tool
}
