% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt-tools.R
\name{g2g_qual_remove_meaningless}
\alias{g2g_qual_remove_meaningless}
\title{Remove meaningless responses from qualitative survey data}
\usage{
g2g_qual_remove_meaningless(response_column)
}
\arguments{
\item{response_column}{A vector (typically a data frame column) containing
qualitative survey responses.}
}
\value{
A vector with NA values and non-substantive responses removed.
}
\description{
This function removes non-substantive responses from a data frame column
representing qualitative responses to a survey. It removes NA values and
common non-substantive responses like "none", "n/a", "nothing", etc.
The filtering is case-insensitive.
}
\examples{
responses <- c("Great class", NA, "none", "N/A", "I loved it", "nothing", "?")
g2g_qual_remove_meaningless(responses)

}
