% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classroom-obs.R
\name{g2g_relabel_yesbut_notreally}
\alias{g2g_relabel_yesbut_notreally}
\title{Relabel a scale that is in math observations to shorten it and so that it matches other scales.}
\usage{
g2g_relabel_yesbut_notreally(response_column)
}
\arguments{
\item{response_column}{Column with scales as responses. Will be \code{response} if \code{tidy_forms_survey()}
is used.}
}
\value{
A character vector of the same length, with the scales transformed.
}
\description{
The scales to relabel go from yes to no, with the ones we are changing starting with 'Yes, but'
and 'Not really'.
}
