% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cleaning.R
\name{g2g_site_information}
\alias{g2g_site_information}
\title{Create key-value pairs of site information}
\usage{
g2g_site_information(site_name, semester, year, subject)
}
\arguments{
\item{site_name}{The name of the site, as a string, and in a style that it would appear in a
presentation. For example, this means that the site name will likely be in title case.}

\item{semester}{The semester, as a string, in title case.}

\item{year}{The year as a integer.}

\item{subject}{The subject in title case.}
}
\value{
A named list that contains the following names: \code{site_name_title}, \code{site_name_lower}, \code{semester_title},
\code{semester_lower}, \code{year}, \code{subject_title}, \code{subject_lower}
}
\description{
Site information is injected into an analysis at multiple points. This function takes the site name,
semester, year, and subject as parameters and returns a named list with standard site information
that can be refered to in a common format.
}
\examples{
g2g_site_information('Bethune', 'Spring', '2022', 'ELA')

}
