% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{g2g_slide_notes}
\alias{g2g_slide_notes}
\title{Add custom notes to Power Point slides}
\usage{
g2g_slide_notes(note_name)
}
\arguments{
\item{note_name}{The names of the predefined note. One of "HE score", "IPG score".}
}
\value{
A string with the note, which can be added to a slide with the \code{notes_text} parameter of
\code{g2g_add_viz_ppt} or \code{g2g_add_table_ppt}.
}
\description{
Predefined custom notes that can be added to Power Point slide notes to help explain the data.
}
