% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{g2g_split_question_stems}
\alias{g2g_split_question_stems}
\title{Create different question stems for stems with more than x number of questions}
\usage{
g2g_split_question_stems(.data, number_questions, grouping_columns)
}
\arguments{
\item{.data}{Input data frame made with \code{gwg_forms_survey_calc_percentages}.}

\item{number_questions}{An integer, the number of questions per plot (per facet)}

\item{grouping_columns}{Columns, as a string vector, that you want to group by when determining whether
the number of questions within the question stem is over the value set by \code{number_questions}.}
}
\value{
The same data frame as before, but question stems containing questions over the value
set by \code{number_questions} contain the word '(continued)' at the start of them.
}
\description{
When creating faceted bar charts that are faceted by question, sometimes there are too many
questions for a single question stem to place on one visualization. This function lets you choose
how many questions should be plotted on one visualization. For questions over this number, it adds the
word '(continue)' to the question stem. Then, the questions will plot to different visualization when
you iterate through stems, creating faceted charts.
}
\details{
The function works with data created by \code{gwg_forms_survey_calc_percentages()}.
}
