% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teacher-survey.R
\name{g2g_teacher_combine_pre_post}
\alias{g2g_teacher_combine_pre_post}
\title{Combine pre and post-training surveys and conduct light cleaning}
\usage{
g2g_teacher_combine_pre_post(pre_training_survey, post_training_survey)
}
\arguments{
\item{pre_training_survey}{The pre-training survey, after running \code{g2g_tidy_forms_survey()}}

\item{post_training_survey}{The post-training survey, after running \code{g2g_tidy_forms_survey()}}
}
\value{
A single data frame with the pre and post-training surveys combined. An additional column is added to
the data called \code{in_survey}. The column will contain "Pre and Post Survey" if the question is in the
pre and post surveys.
}
\description{
Combine the pre and post-training surveys. Then identify question stems in the pre and post training surveys,
and clean up emails and response options.
}
