% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teacher-survey.R
\name{g2g_teacher_shorten_questions}
\alias{g2g_teacher_shorten_questions}
\title{Shorten survey questions}
\usage{
g2g_teacher_shorten_questions(.data)
}
\arguments{
\item{.data}{Teacher survey data set}
}
\value{
A vector with the questions shortened.
}
\description{
Shorten survey question so that they plot easier. Do this by removing text within parenthesis
and shortening long questions.
}
