% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teacher-survey.R
\name{g2g_teacher_viz_pre_post}
\alias{g2g_teacher_viz_pre_post}
\title{Plot pre and post-training comparisons of survey items}
\usage{
g2g_teacher_viz_pre_post(
  .data,
  response_wrap,
  title_wrap,
  reverse_coded = FALSE
)
}
\arguments{
\item{.data}{A data set that contains a single question stem from either the pre or post-training teacher survey.}

\item{response_wrap}{An integer representing the length (number of characters) in each line of the y axis labels.
These will be the questions.}

\item{title_wrap}{An integer representing the length (number of characters) of the title. This is the question stem.}

\item{reverse_coded}{Logical, indicating whether question stem contains reverse coded questions. Defaults to TRUE.}
}
\value{
A ggplot object.
}
\description{
The data plotted should have pre and post-training comparisons with a column called \code{term}
identifying pre and post-training.Reverse coded questions should contain different question stems
than non-reverse coded questions.
}
