% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teacher-survey.R
\name{g2g_teacher_viz_single_survey}
\alias{g2g_teacher_viz_single_survey}
\title{Plot the results of survey items that are only on one survey}
\usage{
g2g_teacher_viz_single_survey(.data, response_wrap, title_wrap)
}
\arguments{
\item{.data}{A data set that contains a single question stem from either the pre or post-training teacher survey.}

\item{response_wrap}{An integer representing the length (number of characters) in each line of the y axis labels.
These will be the questions.}

\item{title_wrap}{An integer representing the length (number of characters) of the title. This is the question stem.}
}
\value{
A ggplot object.
}
\description{
Items only on one survey will not have pre and post-training comparisons.
}
