% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-forms-survey.R
\name{g2g_tidy_forms_survey}
\alias{g2g_tidy_forms_survey}
\title{Transform the pre and post training teacher survey into tidy format (long form)}
\usage{
g2g_tidy_forms_survey(.data, question_columns, grouping_columns = NULL)
}
\arguments{
\item{.data}{The name of an R data frame containing raw teacher survey data.}

\item{question_columns}{An integer or vector of integers representing column numbers for columns contain question answers
that we want to include in the analysis.}

\item{grouping_columns}{An integer or vector of integers representing column numbers of the columns
that provide distinguishing characteristics of respondents that you might want to group on.
This could be a teacher's email address, subjects taught, or demographic information about the teacher.}
}
\value{
A tibble in tidy format where each row represents a single person's response to a single question.
}
\description{
The teacher surveys (pre and post training) are conducted in Google Forms. As a result, their
results are in Google Sheets. In raw form, each question is a column. The function converst the surveys
into a tidy format where each row is a user's response to a specific question.
}
\examples{
teacher_pre_survey |>
   g2g_tidy_forms_survey(8:30, 3)

}
