% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-counts.R
\name{g2g_total_responses}
\alias{g2g_total_responses}
\title{Get the total number of responses for all tools}
\usage{
g2g_total_responses(teacher_responses)
}
\arguments{
\item{teacher_responses}{A data frame created with \code{g2g_get_teacher_names_all_tools()}
listing the tool and all teachers who have entries with the tool.}
}
\description{
Return data frame with the total number of responses for all tools. Data frame
also has a column showing the date / time of the most recent response
}
\keyword{internal}
