% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{g2g_viz_basic_bar}
\alias{g2g_viz_basic_bar}
\title{Create a basic vertical bar chart}
\usage{
g2g_viz_basic_bar(
  .data,
  x_var,
  y_var,
  text_var,
  text_offset = 0,
  fill_color = "gray",
  text_color = "black",
  text_size = 4.11
)
}
\arguments{
\item{.data}{The data set to visualize.}

\item{x_var}{The column name, as a string, of the variable (categorical) on the x axis.}

\item{y_var}{The column name, as a string, of the variable (numeric) on the y axis.}

\item{text_var}{The column name, as a string, of the variable to show as text. This will be the same
number as the variable in \code{y_var}, but users may want to round it or convert it to a percentage.}

\item{text_offset}{Defined how many units the text should offset horizontally from the y variable.
Use negative numbers to show the text within the bar and positive numbers to place the text
above the bar. Defaults to 0, which is the \code{y_var} value.}

\item{fill_color}{The color, as a string or hex number, of the bars for the bar chart.}

\item{text_color}{The color of the text on the bar chart. Default is 'black'.}

\item{text_size}{The size of the text on the bar chart. Default is 4.11.}
}
\description{
Creates a vertical bar chart. This function can be used as-is, but is primarily used to build
more complex plots. For example, \code{g2g_viz_high_expectations()} relies on this function
to create the underlying bar charts.
}
