% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{g2g_viz_checks}
\alias{g2g_viz_checks}
\title{Perform checks for visualizations to ensure entered parameters are correct}
\usage{
g2g_viz_checks(.data, x_var, y_var, fill_var, text_var)
}
\description{
Perform checks for visualizations to ensure entered parameters are correct
}
\keyword{internal}
