% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{g2g_viz_high_expectations}
\alias{g2g_viz_high_expectations}
\title{Bar chart of high expectations score and the percentage of teachers with high expectations}
\usage{
g2g_viz_high_expectations(.data, x_axis, space_between_plots = 50)
}
\arguments{
\item{.data}{Input data frame made with \code{g2g_forms_survey_calc_percentages()}.}

\item{x_axis}{Column name, as a string vector, containing the categories you want to compare.
These will be the x-axis in the plot.}

\item{space_between_plots}{The amount of space, in points ('pt') between plots. Defaults to 40.}
}
\value{
A single plot containing two bar chart plots.
}
\description{
Create two bar chart visualizations in one plot. One containing average high expectations scores and the other
showing the percentage of teachers with high expectations. The data used in the bar charts must
come from \code{g2g_calc_high_expectations_averages()}.
}
