% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{g2g_viz_inst_practice}
\alias{g2g_viz_inst_practice}
\title{Bar chart of instructional practice scores}
\usage{
g2g_viz_inst_practice(.data, x_axis)
}
\arguments{
\item{.data}{Input data frame made with \code{g2g_calc_inst_practices()}.}

\item{x_axis}{Column name, as a string vector, containing the categories you want to compare.
These will be the x-axis in the plot}
}
\value{
A single plot containing instructional practice scores.
}
\description{
Creates a bar chart of instructional practice scores. The data used in the bar charts must
come from \code{g2g_calc_inst_practices()}.
}
