% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{g2g_viz_stacked_bar_percent}
\alias{g2g_viz_stacked_bar_percent}
\title{Stacked horizontal bar chart showing Likert item responses}
\usage{
g2g_viz_stacked_bar_percent(.data, x_var, y_var, fill_var, text_var, color_pal)
}
\arguments{
\item{.data}{The data set to visualize. It must be aggregated results in tidy format. Each row is a
single question and response option ('Agree'), and the aggregate percentage of respondents
- as a decimal (.75) - answering with the given response option.}

\item{x_var}{The x variable name, as a string. This should be numeric and as a decimal between 0 and 1.
It represents the percentage of respondents for the given question and response option.}

\item{y_var}{The x variable name, as a string. This could be questions or a column signifying
pre or post training, with a facet added after this function signifying questions.}

\item{fill_var}{The variable name, as a string, representing the response scales ('Agree').}

\item{text_var}{The variable name, as a string, representing the text to plot over the chart.
This should be numeric and as a decimal between 0 and 1.}

\item{color_pal}{Custom color palette to use. This should be a vector with the values being
the hex codes for the colors and the names being the unique scales from \code{fill_var}}
}
\description{
Creates a stacked horizontal bar chart showing Likert items and responses. Data must be percentages.
}
