% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-counts.R
\name{get_teacher_names_all_tools}
\alias{get_teacher_names_all_tools}
\title{Create data frame of responses for all tools}
\usage{
get_teacher_names_all_tools(list_of_tools)
}
\arguments{
\item{list_of_tools}{A list of all data tools. Each element in the list is a
separate tool. Each tool should also be a list with the following named values:
- 'tool_name': name of the tool ('Teacher Pre-Training Survey')
- 'format': Either 'Google Forms' or 'Qualtrics'
- 'address': For Google Forms, the web address to the Google Sheet containing
the results. For Qualtrics, the survey ID.
- 'name_column': The column name, as a string, where teacher names are located.
Here is an example of the list format:
list(
list(
tool_name = 'ELA Observations',
format = 'Google Forms',
address = "https://googlesheetsurl.com,
name_column = "Teacher's Last Name"
),
list(
tool_name = 'Student Survey',
format = 'Qualtrics',
address = "qualtricssurveyID",
name_column = "teacher_name"
)
)

@returns Data frame of all tools, where each row is a respondent. Columns show
the teacher respondent name and the tool.}
}
\description{
Returns a data frame where each row is a response. Columns are the name of the
respondent and the name of the tool.
}
\keyword{internal}
