% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-counts.R
\name{get_teacher_names_single_tool}
\alias{get_teacher_names_single_tool}
\title{Create data frame of responses for a single tool}
\usage{
get_teacher_names_single_tool(single_tool_list)
}
\description{
Returns a data frame where each row is a response. Columns are the name of the
respondent and the name of the tool.
}
\keyword{internal}
