% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-counts.R
\name{is_date}
\alias{is_date}
\title{Helper function to make sure value can be converted into a date}
\usage{
is_date(date_string, call = rlang::caller_env())
}
\description{
Helper function to make sure value can be converted into a date
}
\keyword{internal}
