% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cleaning.R
\name{scale_order}
\alias{scale_order}
\title{A list containing the order of scales in surveys and observations}
\usage{
scale_order(scale_name)
}
\arguments{
\item{scale_name}{The name of the scale, as a string, or a vector of scale names. Use 'get scale names'
to retrieve a list of possible names.}
}
\value{
A named list where each element contains scales in the proper order. A single vector
of scales can be returned by subsetting on the name of the scale: \code{scale_order()['agree_disagree']}
}
\description{
Returns a named list where each element is a different scale in the proper order. Useful for setting
the order of scales for plotting.
}
\examples{
scale_order('agree_disagree')

}
