% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{system_prompt_teacher}
\alias{system_prompt_teacher}
\title{System prompt for the teacher training program key themes}
\usage{
system_prompt_teacher()
}
\value{
A string containing the system prompt.
}
\description{
System prompt for the teacher training program key themes
}
