% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teacher_pre_survey}
\alias{teacher_pre_survey}
\title{Teacher pre-training survey.}
\format{
A data frame with 17 rows and 33 variables. Each row contains all of a teacher's responses
and is in the same format as Google Sheets creates for Google Forms.
}
\usage{
teacher_pre_survey
}
\description{
A dataset containing example responses to the G2G teacher pre-training survey.
}
\keyword{datasets}
