% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teacher_survey.R
\name{teacher_survey_add_he_metric_colnames}
\alias{teacher_survey_add_he_metric_colnames}
\title{Create tntpmetrics high expectations column name from question.}
\usage{
teacher_survey_add_he_metric_colnames(response_option)
}
\arguments{
\item{response_option}{Vector of high expectations questions. This will be the 'response_option' column
if 'tidy_teacher_survey()' is used.}
}
\description{
Calculating high expectations with the tntpmetrics package requires that each question have a distinct
and pre-specified column name. This function creates the column name based on the high expectations question.
}
\keyword{internal}
