% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teacher_survey.R
\name{teacher_survey_calc_percentages}
\alias{teacher_survey_calc_percentages}
\title{Calculate the percentage and number of responses for each question.}
\usage{
teacher_survey_calc_percentages(.data, grouping_columns = NULL)
}
\arguments{
\item{.data}{A data frame containing the survey responses, created with \code{tidy_teacher_survey}.}

\item{grouping_columns}{A string or vector with the column names of columns that you want to group
the results by. This could include a column containing school names or demographic information.}
}
\value{
A data frame that contains the percentage and number of responses for each response option
and question. Three columns are added, all starting with a period (.):
\itemize{
\item \code{.n_responses}: The number of responses for a given question and response option.
\item \code{.n_total}: The total number of responses for the question.
\item \code{.percent}: The percentage respondents that provided the given response for the question.
}
}
\description{
The teacher survey contains questions that are largely on the 6 point Likert scale. This function
calculates the percentage and number of responses for each question and response option. The data,
\code{.data}, for this function should be created with \code{tidy_teacher_survey}.
}
\examples{
teacher_pre_survey |>
   tidy_teacher_survey(8:30, 3) |>
   teacher_survey_calc_percentages()

}
