% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{user_prompt_teacher}
\alias{user_prompt_teacher}
\title{User prompt for the teacher training program key themes}
\usage{
user_prompt_teacher(survey_question, teacher_responses, num_themes)
}
\arguments{
\item{survey_question}{A string containing the survey question.}

\item{teacher_responses}{A string containing the teacher responses.}

\item{num_themes}{An integer specifying the number of key themes to identify.}
}
\value{
A string containing the user prompt.
}
\description{
User prompt for the teacher training program key themes
}
