% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{viz_fill_barchart}
\alias{viz_fill_barchart}
\title{Plot comparing pre and post training data for data with scales.}
\usage{
viz_fill_barchart(.data, color_pal, x_var, y_var, fill_var)
}
\arguments{
\item{.data}{The dataset used to make the visualization.}

\item{color_pal}{The color palette for the scales. It should be a named vector with the names matching
the scale.}

\item{x_var}{A string representing the column name for the x variable. Will be percent or term (pre / post).
Default is '.percent'.}

\item{y_var}{A string representing the column name for the y variable. Will be percent or term (pre / post).
Default is 'term'.}

\item{fill_var}{A string representing the column name for the fill variable. Will be column with scale.
Default is 'response'.}
}
\value{
A ggplot visualization.
}
\description{
A horizontal bar chart comparing pre and post training data. The x axis is the percentage responses
and the y axis is pre and post training. The plot is faceted by the question. The scale (response column)
is represented by the fill.
}
\details{
Use the function \code{forms_survey_calc_percentages()} to ensure the data is in the proper format.
All default parameter values are designed to with with \code{forms_survey_calc_percentages()}.
}
