% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{viz_pre_post}
\alias{viz_pre_post}
\title{Plot comparing pre and post training data}
\usage{
viz_pre_post(.data, color_pal)
}
\arguments{
\item{.data}{The dataset used to make the visualization}

\item{color_pal}{The color palette for the scales.}
}
\description{
A horizontal bar chart comparing pre and post training data. The x axis is the percentage responses
and the y axis is pre and post training. The plot is faceted by the question.
}
