% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{viz_pre_post_scales}
\alias{viz_pre_post_scales}
\title{Bar chart comparing pre and post with scales.}
\usage{
viz_pre_post_scales(
  .data,
  color_pal,
  x_var = ".percent",
  y_var = "term",
  text_var = ".percent_pretty",
  fill_var = "response",
  facet_var = "response_option",
  facet_str_wrap = 65,
  facet_col = 2
)
}
\arguments{
\item{.data}{The dataset used to make the visualization.}

\item{color_pal}{The color palette for the scales. It should be a named vector with the names matching
the scale.}

\item{x_var}{A string representing the column name for the x variable. Will be percent or term (pre / post).
Default is '.percent'.}

\item{y_var}{A string representing the column name for the y variable. Will be percent or term (pre / post).
Default is 'term'.}

\item{text_var}{A string representing the column name of numeric values to display on plot.}

\item{fill_var}{A string representing the column name for the fill variable. Will be column with scale.
Default is 'response'.}

\item{facet_var}{A string representing the column name for the fill variable. Will be question text.
Default is 'response_option'. Use NULL to avoid any facets.}

\item{facet_str_wrap}{The number of characters in the \code{facet_var} until a new line is created.
Default is 65.}

\item{facet_col}{The number of columns used in the facet. Defautls to 2.}
}
\value{
A ggplot visualization.
}
\description{
Creates a bar chart comparing pre and post surveys with scales. Since scales are used,
the numeric axis is a percentage. Function is designed to work with data produced by
\code{forms_survey_calc_percentages()}, but it can work with other forms of data as well.
All defaults are designed for \code{forms_survey_calc_percentages()}.
}
