% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{g2g_plt_base_theme}
\alias{g2g_plt_base_theme}
\title{Basic theme settings}
\usage{
g2g_plt_base_theme(
  text_font = "Arial",
  center_title = FALSE,
  plot_title_position = "panel",
  font_size = 12
)
}
\arguments{
\item{text_font}{Font for texts. Defaults to 'Segoe UI'.}

\item{center_title}{Should the plot title be centered on the plot? TRUE for yes, FALSE for no.}

\item{plot_title_position}{Takes same values as \code{plot.title.position} in a theme.}

\item{font_size}{Font size to use for text. Title and axis titles will be one point above this font size.
Defaults to 12. 9 is recommended for document, as this is the TNTP font size}
}
\description{
Basic theme settings
}
