% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cleaning.R
\name{g2g_scale_order}
\alias{g2g_scale_order}
\title{Return a vector of scales in the proper order with colors}
\usage{
g2g_scale_order(scale_name)
}
\arguments{
\item{scale_name}{The name of the scale, as a string.}
}
\value{
A a named vector where the values are the scales in the proper order and the names are the
hex codes for colors.
}
\description{
Return a vector of scales in the proper order with colors
}
\section{Scale options}{

\itemize{
\item 'agree_disagree': Strongly Agree, Agree, Somewhat Agree, Somewhat Disagree, Disagree, Strongly Disagree
\item 'knowledge': Excellent Knowledge, Good Knowledge, Some Knowledge, A Little Knowledge, No Knowledge
\item 'how_often': In All or Most Lessons', Often, Sometimes, Rarely, Never
\item 'yes_notyet': Yes, Mostly, Somewhat, Not yet
\item 'yes_but': Yes, 'Yes, But Only in Some Areas', Not Really, No
\item 'almost': Almost Always, Often,  Sometimes, Once in a While, Almost Never
\item 'true': Very True, Mostly True, A Little True, Not True
\item 'high': Extremely High, High, Slightly High, Somewhat High
}
}

\examples{
g2g_scale_order('agree_disagree')

}
