% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt-tools.R
\name{g2g_system_prompt_survey}
\alias{g2g_system_prompt_survey}
\title{Create a system prompt for GPT API}
\usage{
g2g_system_prompt_survey()
}
\value{
A character string representing the system prompt to guide the GPT model in its role and task.
}
\description{
This function creates a system prompt that provides the GPT model with background information
about the task at hand. It sets the role of the model as a social science researcher specializing
in qualitative coding. It also specifies the output format, which is a CSV containing key themes
extracted from the survey responses.
}
