% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt-tools.R
\name{g2g_user_prompt_survey}
\alias{g2g_user_prompt_survey}
\title{Create a user prompt for GPT API}
\usage{
g2g_user_prompt_survey(
  survey_question,
  survey_background,
  survey_responses,
  num_themes = 3,
  custom_language = " "
)
}
\arguments{
\item{survey_question}{A character string representing the survey question.}

\item{survey_background}{Background information on the survey (who took it, purpose, etc.)}

\item{survey_responses}{A character vector representing the survey responses. Each response is a different element in the vector.}

\item{num_themes}{An integer representing the number of key themes to identify (default is 3).}

\item{custom_language}{Additional instructions for the prompt.}
}
\value{
A character string representing the constructed prompt for the GPT model.
This prompt will contain the survey question, context, and responses that need to be summarized.
}
\description{
This function constructs a user prompt to be sent to the GPT API, which includes
the survey question and the survey responses. The goal of the prompt is to ask
the GPT model to identify and summarize the key themes in the qualitative responses
provided by survey participants.
}
