% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{g2g_viz_high_expectations}
\alias{g2g_viz_high_expectations}
\title{Bar chart of high expectations score and the percentage of teachers with high expectations}
\usage{
g2g_viz_high_expectations(
  .data,
  x_axis,
  plots_to_return = "both",
  space_between_plots = 50,
  text_size = 4.21,
  he_scale = "sum",
  ...
)
}
\arguments{
\item{.data}{Input data frame made with \code{g2g_forms_survey_calc_percentages()}.}

\item{x_axis}{Column name, as a string vector, containing the categories you want to compare.
These will be the x-axis in the plot.}

\item{plots_to_return}{String identifying which type of data to return. 'percentages' returns a single plot
showing the percentage of teachers with high expectations, 'scores' returns a single plot showing high
expectations scores, while 'both' returns a single plot containing two visualizations, one with
scores and one with percentages.}

\item{space_between_plots}{The amount of space,in points ('pt') between plots. Defaults to 40.}

\item{text_size}{The size of the text on the bar chart. Default is 4.21.}

\item{he_scale}{The scale in which high expectations scores were calculated.
string, either 'sum' or 'average'. 'sum' is when all HE questions for a teacher
are summed and will fall between 0 and 20. 'average' is when the questions are averaged
and will fall between 0 and 5. Defaults to 'sum'.}

\item{...}{Parameters for \code{g2g_plt_theme()}.}
}
\value{
A single plot containing two bar chart plots.
}
\description{
Create two bar chart visualizations in one plot. One containing average high expectations scores and the other
showing the percentage of teachers with high expectations. The data used in the bar charts must
come from \code{g2g_calc_high_expectations_averages()}.
}
